float32 for WAN

+++ diffusers_nodes_library/common/parameters/diffusion/wan/vace_parameters.py
@@ -50,9 +50,12 @@ class WanVacePipelineParameters(DiffusionPipelineTypePipelineParameters):
 
     def build_pipeline(self) -> diffusers.WanVACEPipeline:
         repo_id, revision = self._model_repo_parameter.get_repo_revision()
+        # Use float32 instead of bfloat16 as a workaround for diffusers bug:
+        # WanVACEPipeline hardcodes video preprocessing to float32,
+        # causing dtype mismatch when VAE is in bfloat16.
         return diffusers.WanVACEPipeline.from_pretrained(
             pretrained_model_name_or_path=repo_id,
             revision=revision,
-            torch_dtype=torch.bfloat16,
+            torch_dtype=torch.float32,
             local_files_only=True,
         )
